requirements.txt
# requirements.txt
pandas
pyarrow
pydicom
presidio-analyzer
presidio-anonymizer
requests
pyyaml 